sdk: reuse an existing GID upon cork create in a limited environment
In some cases, users might want to run `cork create` under a limited
environment, where only limited GIDs are available. Github workflows
CI is one of the examples, since each CI instance is a Docker container
with the latest Ubuntu. In that case, outside of the cork chroot, GID
for group `docker` is `115`, while inside of the chroot, GID for group
`docker` is `233`.

So when `cork create` tries to run a script which tries to run
`useradd -u UID -g 115` inside the chroot, it fails because of the
missing GID 115. So we should actually make the internal script detect
the existing GID 233 to be used for GID of `useradd` command.