kola: skip Docker torcx tests in case of Edge channel
Since the docker 1.12 profile was removed in Edge, many kola tests
started to fail to run, because of the missing docker 1.12 profile.
Ideally we should be able to make a different set of profiles for each
channel. Unfortunately the current Jenkins pipeline is not able to do
that. It takes into account only flatcar-master branch.

To work around the issue, read directly a version from `/etc/os-release`
to detect if the image is Edge. If the semver's minor version is 99,
then it means Edge, so skip the tests.

Note, the code is only a quick-and-dirty approach. In the future, we
should create a generic way to make channel names available for
individual tests.