Port plume/kola to v2/ARM for Azure, remove outdated steps
This uses the v2 API for plume because the classic OS images
are not working with kola and we also migrated the storage
account.
The prerelease step is also simplified because the
image is not necessary for the followed release step as it
directly works with page blobs.
For kola tests we better create an image where needed.
The "pre-publish" container is not used anymore.
Instead of "sharing" the image we now create a SAS URL
which we need to specify in the SKU. This step remains
manual for now.
In addition, there are new command line flags for testing plume.
For kola the options --azure-image-file and --azure-blob-url
were added to directly upload a file into the temporary resource
group or to copy a blob from another storage location into the
temporary resource group (the blob may be uploaded before via
"plume pre-release" or "ore azure upload-blob-arm".