kola list: Platform/distro filters with glob patterns
The "kola run" command accepts a glob pattern to limit which tests
are run.
To test a pattern before running a test, add a filter option to
"kola list" to filter the list of known tests by platform and distro,
and if wanted with a glob pattern to see its effect.