kola run: Accept a list of glob patterns
Glob patterns are matching like the shell expands a glob pattern.
This does not allow to test against two names that don't share a
substring. On the shell this is solved by having two glob arguments.
Allow multiple glob patterns to be specified as arguments to handle
the cases where the tests don't share a substring but we want to run
exactly those tests.