kola: Add YAML description with error message to TAP file
The TAP file is not useful if the console logs are lost because the
TAP files does not contain the error message.
Add a YAML description to the TAP file as understood by the Jenkins
TAP plugin. For now it only contains the test's error message as value
for the "Error" entry.