plume: Remove duplicate architecture tag from AWS image name
The architecture tag was first added to the image before the -hvm suffix
and then later added to the image metadata.
Since the metadata is used for the image name, we ended up with
adding the architecture tag twice for the pending release:
Flatcar-edge-2466.99.0-arm64-arm64-hvm
Do not append the architecture tag before the -hvm suffix.