kola: increase cgroup memory limit to 30MB
Since runc 1.0-rc7 or newer, memory consumption of runc has increased,
so Docker has not been able to run containers with cgroup memory limit
> 10MB. The memory consumption issues were improved since 1.0-rc8, but
not completely. That's why some docker tests have failed with the
following messages.

```
--- FAIL: docker.base/resources (34.69s)
        cluster.go:117: /run/torcx/bin/docker: Error response from daemon: cannot start a stopped process: unknown.
        docker.go:302: [0] failed to run "docker run --rm --memory=20m --memory-swap=20m sleep sleep 0.2": output: "" status: "Process exited with status 125"
```

We need to increase the limit from to 30MB, to make the tests work
again.