Switch to Hyperkube v1.14+.
Starting from Docker 19.03, we need a newer version of Kubernetes or our
deployment doesn't work. So we switch to testing newer k8s versions.

This includes updating quite a few Kubernetes flags that were deprecated
since the old version we were using, to use the new names and formats.

Supporting different Kubernetes versions also means adding some
conditional behavior for commands that changed along the way.

This commit also switches to the hyperkube images provided by gcr.io
instead of the ones provided by quay.io, as those were discontinued. To
get this to work we need to prepend the ACI with docker:// and set the
--insecure-options=image option (See coreos/coreos-kubernetes#583)

It also updates the environment variables in the service files to the
non-deprecated ones wanted by kubelet-wrapper.

And it adds a few plog messages while setting up to make the kola output
more helpful.