Packet: Always report error when provisioning failed
The retry loop relies on having access to the last error. This error value
was shadowed in some cases which caused the error not be be logged and the
function to report a nil error when the maximal number of retries was reached.

Never shadow the error variable with local error variables in for/if scopes so
that the error value is assigned to the functions error variable.