ESX: Add kola option to work without a base VM
Cloning from a base VM to spare the image upload is not supported by
the free version of ESXi. It is also useful for "kola spawn" to work
without a base VM.
Add an option to create a VM directly from an uploaded image.
The userdata is set over guestinfo variables instead of modifying the
default values of the OVF metadata. Use the new generic ignition.config.*
variables instead of coreos.config.*.