kola/tests/locksmith/locksmith: Make coreos.locksmith.tls more robust
On VMware ESXi there were reproduceable test failures with kola due to
the certificate date ("certificate is not yet valid") but these
failures were not reproducible by executing the same commands manually.
Cover the failing command in a rety loop because I suspect that the
system time got slightly adjusted or something similar in between
which causes the certificate to be invalid for a short time.