Azure platform: fix getting the console output
The regex patterns to process the URL were assuming that the test names
would all start with "kola-", which is not true when running Jenkins
tests. This caused the container and blobName variables to be empty,
making it impossible to get the console output.

On top of this, the function that reads the data was always returning
an error, which meant that the data was never written, even if it could
be obtained successfully.

This change fixes the regex patterns, adds a lot more error handling,
and returns non-error when the function succeeds