cmd/plume: Add a prune command to clean up developer images
When running developer builds, plume uploads prerelease images to both
Azure and AWS. This change adds a new "prune" subcommand for plume, to
get rid of artifacts uploaded to old builds.

It also adds or improves logging messages on functions. This extra
logging was needed for debugging while developing the code.