azure: use one Resource Group for all tests
Each kola cluster created and destroyed its own Resource Group which
took 2-3 minutes.
Reuse the image Resource Group if it was created for the flight and
run all tests in it. To keep the amount of left over VM network objects
small, the IP addresses and NICs are deleted when an instance is
deleted - this could also be removed again if we see that deleting
these objects takes a lot of time.