cork: Set SDK version in the manifest version.txt
The SDK version parameter for "cork create" and "cork update" was used
to download a specific SDK tar ball but the SDK version entry in
version.txt stayed on the original value of the manifest. This caused
"cork update" or a later update_chroot step to not find the right
binary packages.
Set the correct SDK version value in the version.txt file.