Clarify SSH check error message for new machines
The CheckMachine functions waits until an SSH connection runs
"systemctl is-system-running" with an output like "running" or
"degraded". However, the error message always suggested that SSH
would be unable to connect because it only stated "ssh unreachable".
Extend the error message to include the case of the system not being
ready and give more context for the "systemctl is-system-running"
output if it failed.