platform/qemu: Retry the mount operation
For some unknown reason, sometimes the loop devices are generated
twice, with a brief moment in between where the devices aren't there. If
the mount call happens to be issued in that window, it will fail. So,
retry the mount operation a few times before actually failing.

Also, while investigating this, I've refactored the other Retry function
to avoid a condition where it could mask errors different than
os.IsNotExist.