Packet: Set OEM ID for PXE boot network setup
Without any OEM information the default DHCP setup for one of the
interfaces works ok and is required to work inside the initramfs
but later after the initramfs the network setup provided by the
metadata server should be used. This is what is currently used in the
real PXE installation processes and should be covered by our CI. It is
also needed as soon as there is a change that assumes the network setup
according to the metadata server is done after the initramfs.