Add ability to make GCE images public, use it in ore and plume
This change is based on f5d422d365add04c4e588570afe531ab9e4e5f52 from
https://github.com/coreos/coreos-assembler/ (coreos/coreos-assembler#1610)

It adds support for marking images with public ACLs when they are
created, and uses this functionality from ore upload, ore create-image
and plume release. In the ore subcommands, this is gated with a --public
flag, in the plume subcommand, the image is always made public.