cmd/cork,sdk: introduce --repo-branch option to cork
Introduce a new option `--repo-branch` to `cork [create/update]`.
The option allows users to specify an arbitrary branch name passed to
the `repo init` command.

Also set the default branch name of repo to `maint`, until the upstream
repo >= v2.10 could be available in ordinary SDK environments. That is
to avoid incompatibility issue of an old repo tool not being able to
work with the default `stable` branch of the repo tool,
https://gerrit.googlesource.com/git-repo/+/refs/heads/stable,
which does not support python2 any more.
OTOH its `maint` branch still supports python2.

In the long term, we should update `dev-vcs/repo` in Flatcar SDK to
v2.10 with python3, and set the default branch back to `stable`.