cmd/cork: Allow to apply a patch on top of the manifest references
Sometimes the git branches the manifest refers to should not be
changed but still a modification introduced. This can now be done
in the form of a patch file for each of the three repositories that
define what gets build. A single patch file is taken but "git am"
is able to process the concatenation of multiple files well. The
three-way merge option is set to be able to use slightly outdated
patches as long as they get merged automatically by git.