dnsmasq: add veth generation and configuration
this new topology allows QEMU VM to access internet. This allows
tests with RequiresInternetAccess flag to now be run.

on the implementation PoV: we basically create a virtual ethernet
pair generated from a seed (a TCP socket listener in the root namespace)
Then we configure routing and firewall to use this pair in order to
forward packet until root network namespace