platform: allow CLC templating for dynamic IP address insertion
The tests that require coreos-metadata and CLC templating could not
run because for QEMU only a static replacement was done before the
CLC configuration was rendered.
Add support for the template variables by treating QEMU as a custom
platform for the transpiler, making use of a coreos-metadata service
that has the IP addresses hardcoded.