tests/cl.filesystem: skip "/run", fix test failure on arm64
On arm64, we have a `/run/udev/data/+platform:Fixed MDIO bus.0`, which
gets flagged due to the presence of spaces in the filename. Since /run
is an ephemeral directory and we skip other directories (/proc, /sys)
for that reason, it seems like a good idea to skip /run as well.