system: fix check for errror when checking for symlink
The check of 'err' against nil was inverted and so the whole function
would return false as long as path exists, regardless of whether it is
a symlink or not.
IsSymlink() is used to check if /dev/shm is a symlink, which is the
case atleast in WSL2. It needs to behave correctly to trigger the
proper fallbacks.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>