sdk: correct setting sticky bit for /dev/shm in fallback
The go stdlib seems to have special handling for the sticky bit in
mkidr/chmod, and doesn't accept it as part of the octal mask but
requires explicitly or'ing in the correct constant. Without this commit,
the /dev/shm fallback path (triggered in WSL2) would not have the
sticky bit set.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>