kola/tests/verity: fall back to expected dm-verity offset
When e2size can't find the size it returns an error exit code and
stdout stays empty, failing the test. However, the hash offset is
actually fixed because the GPT disk layout has to stay the same in
Flatcar Container Linux as the partition contents are swapped out
when updating. In case another filesystem like btrfs is used, e2size
doesn't work and it makes sense to fall back to the only value which
is supported in general.