kola/tests/verity: skip the corruption test for btrfs
The corruption test corrupts a particular file and checks that read
syscalls for this file fail. With btrfs is is not easy to find the
right block to corrupt. It is also currently discussed to issue a
kernel panic on corruption because failed read syscalls still allow to
alter the system behavior and are harder to spot in logs.
For now, just skip the corruption test for btrfs and later we can
update it to check that kernel panics work when any corruption is
detected, not only a particular file.