Expect kernel panic when dm-verity detects corruption
The dm-verity behavior was to only let some syscalls fail but it let
the system run in a broken state. By enabling kernel panics on
corruption detection we get a more reasonable behavior. This change is
done in https://github.com/kinvolk/bootengine/pull/26
To support the new behavior the test that do corrupt on purpose need to
be flagged to be excluded from the kernel panic checks. The corruption
test also shouldn't test for a failed syscall anymore but expect the
kernel panic to render the machine unreachable.

Implement the test flag to ignore kernel panics on the console and
modify the corruption test to fit the new behavior, which also makes
the test easier because it doesn't need to find the exact file to
corrupt and we can destroy the filesystem metadata and then try to read
any file.
Since there is currently no way to specify which release version a test
requires, all channels have to be skipped for now until shortly before
the release of each channel is built.