kola: support nightly version in version comparions
The nightly version is the build date but the build ID has the branch
reference. To support the version comparion for nightly builds we need
to generate a version for the "main" branch and the major version
branches that is higher than the current release version.

Use 999999.99.99 as version in the semantic version checks for nightly
builds from the "main" branch or custom dev builds and use MAJOR.99.99
as version for nightly builds from the release branch.