platform/qemu: retry if OEM btrfs filesystem is in use
The OEM partition can't be mounted more than once when it's a btrfs
filesystem, because having the same UUID is reserved for RAID usage.
When two kola instances run at the same time there is a possibility
for a race when both shortly want to use the OEM partition.

Use a conditional retry waiter to wait for exclusive access, giving
up after 10 minutes.