kola/tests/misc/verity: add recursive list to provoke panic
On amd64 the panic was triggered by the 'cat' execution as expected
but on arm64 through QEMU emulation the kernel panic took a while to
be triggered. It seems the most reliable way is to walk the filesystem.

Add a recursive list to provoke the panic on arm64 if it didn't get
triggered by the simple 'cat' file reading.