platfor/api/azure: use azcopy for blob copy if available
The Azure SDK does server side copy but does not have PageBlob related
optimizations, so the whole blob is copied. This causes the copy to be slower
than necessary (30GB vs 1GB). If we have azcopy on the local machien, use it to
copy the blob because this is much faster.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>