Jenkinsfile: also build arm64 artifacts
With this change we no longer need the job parameter so it has been removed.
The arm64 build requires disabling 'CGO_ENABLED' as otherwise we would need a
cross-compiler in the build container. The arm64 artifacts are now archived at
bin.arm64. The build has also been switched to go1.16, this is what we are
using in the Flatcar SDK as well.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>