kola/test: use clc to set `enable_v2`
In 12b5a96ee4099cc3c2dbe4af35ebbfc55b13868a, we enabled the V2 support
for `etcd`. It means that `etcd` starts with support for both datastore
version.

Support has been enabled through a `systemd` dropin - this one will
generate a dropin under the `etcd-member.service` unit but
`container-linux-config` will generate a dropin under
`etcd-member.service` name too.

This is not an issue with `ignition/v2` but it is with `ignition/v3`
(see: https://github.com/kinvolk/ignition/blob/main/docs/configuration-v3_0.md).

So we need to add the `enable_v2` direcly into the `clc` - `enable_v2`
is present in `clc` from `config.Types.Etcd3_2` but by default it
targets `3.0.0` so we need to pin the version to the current tested one
(`3.5.0`) to use `enable_v2` and the correct image tag.

Signed-off-by: Mathieu Tortuyaux <mtortuyaux@microsoft.com>