kola/misc/raid: change CLC to raw ignition config
`size` has been deprecated to `sizeMiB` from `ignition-2.3` and this
parameter is not translatable. See the discussion in here: https://github.com/coreos/ign-converter/pull/29

Since CLC is converted to `ignition-2.2` we can't rely on this to pass
the config so we directly use raw ignition 2.3.0 to provision the
instance.

This test only runs on `qemu` so there is no side effect to pass from
`clc` to `ignition`.

Signed-off-by: Mathieu Tortuyaux <mtortuyaux@microsoft.com>