kola/harness: find right AWS CPU architecture for kolet/skipping tests
When scpKolet is called it needs to know which board binary to copy.
With the default amd64 we get this error on AWS arm64 instances:
  bash: line 1: ./kolet: cannot execute binary file: Exec format error
The same applies to isAllowed(architecture(platform) which is used to
skip tests that should only run on amd64, like cl.update.grubnop.

Find the right architecture when running on AWS.