RenderUserData: only incject SELinux enforcing if there is userdata
Some tests use machines without userdata and in this case there is no
way to inject it and the test should rather use a minimal dummy
Ignition config to enforce SELinux if it's relevant.

Don't try to inject SELinux enforcement in case no userdata is present
because the function panics in this case.