plume: Use the public bucket proxy for anonymous access
The image download from the bucket is not possible anymore directly but
needs to be done through the HTTP proxy.

Create a limited bucket object for anonymous access, which can only
download particular files and has no listing or upload ability.
Calling code needs to make sure it only tries uploads or directory
listings when the bucket object is using a proper gs:// URL and that
it configured credentials before trying to upload.