Equinix Metal: Don't wait for /dev/sda when installing
Some instance types like the c3.large.arm have only a single NVMe disk
and the installer waits endlessly for /dev/sda while it's not required.

Remove the waiting for /dev/sda from the installer service unit and
make it be restarted in case of errors, like not having found a
suitable disk.