kola/tests/etcd: add `name: ` to one CLC
Using both command line option and environment variable is forbidden
since `etcd/v3`.

It happens quite a lot that users override default `etcd` name
(`machine-id`) using `CLC` through the `name: ` key.
In this case it fails because we will have both `ETCD_NAME`
and `--name` in the runtime.

This patch overrides the name in CLC in order to reproduce this behavior
and prevent regression in the next release.

Signed-off-by: Mathieu Tortuyaux <mtortuyaux@microsoft.com>