platform/api/azure: add support for gallery images
GalleryImages can be created from a blob url, and can be republished on Azure
marketplace in a simpler way than a vhd blob. This uses an ARM template as the
azure-sdk-for-go currently does not support all parameters that we need to use.

The gallery support is not enabled by default as there may be differences in
semantics that we want to figure out before it becomes the default path.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>