kola: Ignore dead links in /run/udev/links
Previously the entries in the directory were like `rtc/c252:0`, where
`c252:0` was just an empty file. Starting from systemd 250 it became a
dead symlink which points to `-100:/dev/rtc0`. This has caused the
`cl.filesystem` test to fail. To fix it, we add another entry to the
ignore list.