platform/platform: reduce sshRetries to 60
The way util.Retry works is, it attempts an operation multiple times, with a
delay between each attempt. The current values for ssh operations are 120
attempts at 10s delays between attempts. Unfortunately, the ssh operation will
itself block with a connect timeout, adding to the delay between attempts. So
the operation time is not bounded to 1200s, but can be much longer.  As an
example, on AWS tests failing to establish a journal connection have in some
cases taken 6000s, which is excessive. We will rework the retry logic but for
now, it is a good idea to reduce the number of retries.