kubeadm: extend tests following the CNI
kubernetes tests should be agnostic from the CNI, but some specific CNI
features need to be tested.

with this commit, we introduce a mechanism to extend the kubeadm
test suite by adding `extraTests`.
An `extraTest` is bound to the CNI, this aims to be used to extend CNI
coverage.

`extraTest` are ran from the `controller` node.

Signed-off-by: Mathieu Tortuyaux <mtortuyaux@microsoft.com>