azure: add support for using existing virtual network
The user can now pass the name of an external vnet/subnet to use when creating
instances. If the subnet name is missing, 'default' is assumed. The validation
of the parameter is performed after creating resource group and storage account
so in case of failure we need to destroy them. The easiest way to accomplish
that is by calling the Flight or Cluster destroy methods.