platform/azure: cache Subnet struct after creation
When preparing a flight, network resources are created. Later when creating an
instance, we need to lookup the resources again in order to pass the Subnet
struct to the NIC creation function. Instead of performing the lookup every
time an instance is created, store it in the flight and cluster structs.