platform/api/azure: support accessing instances on private ip addresses
Kola now supports creating instances inside an existing virtual network. When
combined with Azure VPN gateway, or when running tests from within that virtual
network, it is possible to access instances via private ip addresses. We signal
this by setting an empty public IP name, which shortcuts the lookup of the
public ip address and causes it to return the private IP.

We still need to give the instances a public IP for outbound networking.