platform/api/equinixmetal: wipe disks before installing
On EM disk names are not stable and may change across reboots. This is
generally not a problem, but since we started reusing machines, it is
much more likely that there will be a leftover flatcar installation on
one disk and we are now installing to the second. This messes with some
tests because partitions from both disks may be mixed up when mounting
by label. Fix this by wiping the partition tables from disks on
reinstallation.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>