kola/tests/misc/grub.go: Try to fix grubnop test
The mask clause in coreos-cloudinit was masking the unit and then
telling systemd to reload. With that, systemd noticed that update
engine is masked so systemd killed it, and for some reason marked it
as failed.

Current approach is to have a unit that conflicts with update engine
and masks it. That way, systemd should stop update engine without
marking it as failed and then mask it.

Co-authored-by: Kai Lüke <kailuke@microsoft.com>