cl.filesystem/deadlinks: ignore factory /etc in /usr/share/flatcar/etc
The copy of the factory /etc in /usr/share/flatcar/etc may have dead
links because the relative linking assumes to be under /TOPDIR
(https://github.com/flatcar-linux/scripts/pull/264).
Ignore the folder in the check for dead links, we still check the
resulting /etc hierarchy created based on it.