.github: Publish mantle container image
A container image addresses the needs of running arm64 kola tests
without QEMU emulation but with KVM on arm64 hardware because there is
no native Flatcar SDK for arm64 yet, and it also allows us to have a
single place that provides a particular version of the mantle binaries
that is more lightweight than the full SDK where rebuilding the ebuild
package is currently needed.
The container image is published for the main branch but uses the git
SHA commit ID as image tag which should be updated in the "scripts"
repository.
There is no garbage collection yet to remove images created by PR
events.