test/ignition: add networkd test configuration
This commit creates a dummy network interface with a dummy IP assigned.
With recent Ignition, `networkd` is no more used and should be
translated to `files` section.

This test ensures that we keep the `networkd` behavior by translating it with
`ign-converter`.

This test lives with an existing test (both are independant) to not add a new test
and it will run only on QEMU.

Signed-off-by: Mathieu Tortuyaux <mtortuyaux@microsoft.com>