api/equinixmetal: implement the ssh storage
this defines a ssh storage implementing the Storage interface. This
implementation can be used if the `--equinixmetal-storage-url` used the
following schemes:
* ssh
* ssh+http
* ssh+https

In the Upload method, a SSH connection is made and the file is uploaded
on the document root of the remote server.

The returned URL is built from the storage URL scheme (http or https)
and the document root.

It's up to the operator to deploy and configure Webserver and SSH setup.

Signed-off-by: Mathieu Tortuyaux <mtortuyaux@microsoft.com>