kola/tests: Limit generic tests to QEMU
The number of tests causes the test time to be many hours on some of
the platforms and also costs money and energy while many tests are not
really adding much value when being tested on each of the platforms.
Some very generic tests already were limited to QEMU which is a good
solution for now, even if that means that the execution of all possible
tests on a platform now requires a source code change because we don't
have a flag to allow running additional tests when wanted.
Limit the execution of generic tests to QEMU. These generic tests are
not expected to have a different result on a cloud environment unless
something fundamental is at error. For these fundamental things we
rely on tests that still get executed on each cloud, e.g., to test
Ignition and networking, but we also rely on end-to-end tests such as
the kubeadm test. Each test now has a comment that tells whether the
execution on all clouds is wanted or not and why. The hope is that with
focused tests we don't waste our time to wait for unnecessary tests or
even have to restart them multiple times because they are flaky. Also,
with a reduced list of tests we can easily add new ones that we think
are missing but didn't realize before because there were to many, or
didn't want to add because the test time was too high already.

See https://github.com/flatcar-linux/Flatcar/issues/654