api/equinixmetal: Exclude devices with block extended major
I don't know if NVMe disks are a recent addition to s3.xlarge.x86
instances, or maybe those disks got shrunk, but the flatcar-install
script was picking one of them to install flatcar on. For some reason,
the boot agent is not able to boot from them, maybe there is some
additional setup needed, which is missing. Fortunately the script
already had an option of excluding devices by major numbers, so use
this functionality. That way, the script installs the OS on one of
`/dev/sdX` disks, which are bootable.

For device majors see the kernel documentation at:
https://www.kernel.org/doc/Documentation/admin-guide/devices.txt