platform/api/aws/images: return snapshotID together with found image
Instead of searching for snapshots afterwards, we can extract the snapshot id
from image metadata. This allows us to correctly identify the snapshots we want
to delete, which will solve an issue with some older releases where -hvm and
normal image variants have the same tag:Name and so two snapshots are found and
can't be pruned.