Dockerfile: Use cross-compilation instead of binary translation
To compile mantle for arm64 we relied on binary translation on the
amd64 GitHub runner. This turned out to be quite slow, in total we
needed 1 hour for the combined amd64+arm64 container image build.
Explicitly use the amd64 docker image in the builder and do a
cross-compilation to speed the build up for arm64. There are no
changes for amd64 because we keep CGO but for arm64 we have to
disable it. This will only affect kola arm64 qemu tests. We may
later consider to disable CGO always for consistency.