kola/harness: run `SkipFunc` when version is available
`SkipFunc` is called from `FilterTests` and `FilterTests` is called
twice: before and after spining-up a test instance to get the version.

We need to assert to be in the second case, where the version is
actually defined - otherwise it simply ignores the test from the start.

Notified from the VMWare kubeadm that did not run on main nightly.

Signed-off-by: Mathieu Tortuyaux <mtortuyaux@microsoft.com>