kola: add --force-flatcar-key flag
this flag will inject the Flatcar production key to verify the signed
payload in the instance to update.

It can be useful to verify actual signed payload and not the payload
generated by the CI (which uses the development key).

It defaults to the current behavior: inject the development key.

Signed-off-by: Mathieu Tortuyaux <mtortuyaux@microsoft.com>