platform/azure: fetch marketplace image plan info from endpoint
This is not a frequently used codepath, but when a user specifies an image
through publisher/offer/sku/version we may need to pass plan information when
creating the vm. So far we made 2 assumptions:

 a) plan info = image reference
 b) plan info is always needed with an image reference

These don't hold up now that we have an arm64 corevm offer which fails to
provision when plan information is passed. Fix this by querying plan info at
runtime, and passing it to the vm create call.