kola: Add new cl.sysext.boot test for selection of active OEM sysext
For OEM sysext images and optional Flatcar extensions we need on-boot
selection of the active sysext image and possible moving between the
rootfs and the OEM partition. This is introduced in
https://github.com/flatcar/bootengine/pull/50
Add a new test to check that the boot logic works as expected by
preparing a QEMU machine to cover the relevant conditions. Since the
systemd units in the initrd and the mounting is different depending on
whether Ignition runs and itself interacts with the OEM partition, we
have two machines, one without Ignition and one with because we anyway
need to aquire the full version string first to be able to create the
Ignition config with the right entries.