kola/harness: Add more checks for common failures
When testing changes with kola and the tests do not cover the changed
area we either have a booting system (that passes the tests) or don't.
With additional generic checks we can catch introduced errors in an
automated way and the manual booting to inspect a boot failure can also
be saved if kola can hint at errors. Examples are broken changes in the
systemd unit setup, when services start to fail, or when the disk is
corrupted (be it due to a kernel bug, broken hardware, or a cloud
provider image problem).
Add more checks for the above failures. If a test provokes one of these
failures, the test and the pattern will have to be marked with a flag
that skips the patterns for those tests only (this is what we already
do for the existing patterns).