kola/tests: Add test for /etc overlay upcopy prevention and cleanup
The /etc overlayfs mount is there to be able to update the lowerdir
file and this only works if we have no unwanted duplicates in the
upperdir (i.e., the /etc folder).
Add tests to check against upcopies caused by, e.g., systemd-tmpfiles
and to check that duplicates get removed on reboot
(introduced in https://github.com/flatcar/coreos-overlay/pull/2482).