kola: add DefaultUser register parameter
this parameter allows to override the default user ('core') when a test
is registered.

This user will have the same SSH keys as the 'core' user (the one created
by Mantle and the one provided by the CLI if any) and it will be added to the
'sudo' group by default.

Other groups and other properties can be directly added from the
initial provisioning defined in the test.

This user will be used for SSH operations and it can be used for
debugging too.

Signed-off-by: Mathieu Tortuyaux <mtortuyaux@microsoft.com>