kola: Add new OEM sysext update and migration test
The A/B updated OEM systemd-sysext image gets activated when both /usr
partitions have updated to a version that requires a sysext image. The
old OEM contents get cleaned in this migration that happens on the boot
after the final update.
Test this migration and update logic by starting from an old image that
isn't migrated, and emulate an OEM setup that requires an OEM sysext
image now. The update first happens with the inbuilt Omaha kolet server
which uses the fallback logic but then flatcar-update is used for the
second update to supply the OEM payload directly.