kola: Skip cl.sysext.fallbackdownload for release builds
We currently have no way to switch the update pub key in the initrd to
be a dev key (or vice versa). Thus the test cl.sysext.fallbackdownload
failed because based on the key either the update server or bincache is
tried.
Skip cl.sysext.fallbackdownload for release builds at least temporarily.
Later we might use a kernel cmdline setting or similar to switch the
key to a dev key.