kola: Disable the public update server by default
This might have been an oversight: for release tests we used the public
Nebraska server. This was already disabled for FCOS but not for the
original CoreOS/Flatcar CL.
Disable the public update server by default. The update tests overwrite
it with their specific setting. In case a future test needs the default
the usual test flag escape hatch in the form of NoDisableUpdates is
added.