tests/docker.go: explicitly configure btrfs driver in daemon.json
This change adds an explicit brtfs driver request in
/etc/docker/daemon.json to un-break the docker btrfs test for docker 23
and above. The explicit storage driver configuration does not break
docker 20 and older; it works fine if btrfs storage is provided (which
it is in the test).

Starting with docker 23, overlay2 is the default also for btrfs
filesystems, and the btrfs driver is opt-in. See
https://docs.docker.com/engine/release-notes/23.0/#bug-fixes-and-enhancements-6
and
https://github.com/moby/moby/pull/42661

Signed-off-by: Thilo Fromm <thilofromm@microsoft.com>